+++ .gitignore
@@ -1,1 +1,3 @@
 venv/
+YOLO*.pt
+yolo_weights/
